add user authentication
