--- src/main/java/org/clx/library/model/Staff.java
@@ -1,17 +1,0 @@
-package org.clx.library.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@PrimaryKeyJoinColumn(name = "user_id") // Join on user_id in the staff table
-public class Staff extends User {
-    private String department;
-    private String position;
-
-}
