--- src/main/java/org/clx/library/model/Student.java
@@ -1,17 +1,0 @@
-package org.clx.library.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@PrimaryKeyJoinColumn(name = "user_id") // Join on user_id in the student table
-public class Student extends User {
-    private String course;
-    private String year;
-
-}
