--- src/main/java/org/clx/library/repository/StaffRepository.java
@@ -1,9 +1,0 @@
-package org.clx.library.repository;
-
-
-import org.clx.library.model.Staff;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StaffRepository extends JpaRepository<Staff, Integer> {
-}
-
