--- src/main/java/org/clx/library/repository/StudentRepository.java
@@ -1,7 +1,0 @@
-package org.clx.library.repository;
-
-import org.clx.library.model.Student;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StudentRepository extends JpaRepository<Student, Integer> {
-}
