--- src/main/java/org/clx/library/response/ApiResponse.java
@@ -1,14 +1,0 @@
-package org.clx.library.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ApiResponse {
-    private String message;
-    private boolean status;
-}
-
