+++ src/main/java/org/clx/library/user/config/JwtValidator.java
@@ -1,4 +1,4 @@
-package org.clx.library.config;
+package org.clx.library.user.config;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
