+++ src/main/java/org/clx/library/user/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package org.clx.library.config;
+package org.clx.library.user.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
