+++ src/main/java/org/clx/library/user/exception/EmailAlreadyUsedException.java
@@ -1,4 +1,4 @@
-package org.clx.library.exception;
+package org.clx.library.user.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
