+++ src/main/java/org/clx/library/user/request/LoginRequest.java
@@ -1,0 +1,14 @@
+package org.clx.library.user.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginRequest {
+    private String email;
+    private String password;
+}
+
