+++ src/main/java/org/clx/library/user/response/ApiResponse.java
@@ -1,14 +1,14 @@
-package org.clx.library.response;
+package org.clx.library.user.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
-@AllArgsConstructor
 @NoArgsConstructor
-public class AuthResponse {
-    private String token;
+@AllArgsConstructor
+public class ApiResponse {
     private String message;
-
+    private boolean status;
 }
+
