+++ src/main/java/org/clx/library/user/response/AuthResponse.java
@@ -1,4 +1,4 @@
-package org.clx.library.request;
+package org.clx.library.user.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,8 +7,8 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class LoginRequest {
-    private String email;
-    private String password;
-}
+public class AuthResponse {
+    private String token;
+    private String message;
 
+}
