+++ src/main/java/org/clx/library/user/service/serviceimpl/UserServiceImpl.java
@@ -1,14 +1,12 @@
-package org.clx.library.service.serviceimpl;
+package org.clx.library.user.service.serviceimpl;
 
 import lombok.RequiredArgsConstructor;
-import org.clx.library.config.JwtProvider;
-import org.clx.library.model.Staff;
-import org.clx.library.model.Student;
-import org.clx.library.model.User;
-import org.clx.library.repository.UserRepository;
-import org.clx.library.service.UserService;
+import org.clx.library.user.config.JwtProvider;
+import org.clx.library.user.model.User;
+import org.clx.library.user.repository.UserRepository;
+import org.clx.library.user.service.UserService;
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory; // Import SLF4J Logger
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -22,19 +20,7 @@ public class UserServiceImpl implements UserService {
     public User registerUser(User user) {
         logger.info("Registering user: {}", user.getEmail());
 
-        User newUser;
-
-        if (user instanceof Student) {
-            newUser = new Student();
-            ((Student) newUser).setCourse(((Student) user).getCourse());
-            ((Student) newUser).setYear(((Student) user).getYear());
-        } else if (user instanceof Staff) {
-            newUser = new Staff();
-            ((Staff) newUser).setDepartment(((Staff) user).getDepartment());
-            ((Staff) newUser).setPosition(((Staff) user).getPosition());
-        } else {
-            newUser = new User();
-        }
+        User newUser = new User();
 
         newUser.setId(user.getId());
         newUser.setName(user.getName());
@@ -54,13 +40,13 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User findUserByJwt(String jwt) {
-        logger.info("Finding user by JWT: {}", jwt); // Log the JWT used for finding a user
+        logger.info("Finding user by JWT: {}", jwt);
         String email = JwtProvider.getEmailFromJwtToken(jwt);
         User user = userRepository.findByEmail(email);
         if (user != null) {
-            logger.info("User found: {}", user.getEmail()); // Log found user
+            logger.info("User found: {}", user.getEmail());
         } else {
-            logger.warn("No user found for email: {}", email); // Log warning if no user found
+            logger.warn("No user found for email: {}", email);
         }
         return user;
     }
