+++ src/main/resources/application.properties
@@ -1,16 +1,16 @@
 spring.application.name=library
-#spring.datasource.url= jdbc:mysql://localhost:3306/clx
-#spring.datasource.username= admin
-#spring.datasource.password= XDuce@123
-#spring.datasource.driver-class-name= com.mysql.cj.jdbc.Driver
-#spring.jpa.hibernate.ddl-auto=update
-#spring.jpa.show-sql=true
-#spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
-
-
-spring.datasource.url=jdbc:postgresql://localhost:5432/clx
-spring.datasource.username=postgres
-spring.datasource.password=Admin@123
+spring.datasource.url= jdbc:mysql://localhost:3306/clx
+spring.datasource.username= admin
+spring.datasource.password= XDuce@123
+spring.datasource.driver-class-name= com.mysql.cj.jdbc.Driver
 spring.jpa.hibernate.ddl-auto=update
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.hibernate.show-sql= true
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
+
+
+#spring.datasource.url=jdbc:postgresql://localhost:5432/clx
+#spring.datasource.username=postgres
+#spring.datasource.password=Admin@123
+#spring.jpa.hibernate.ddl-auto=update
+#spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+#spring.jpa.hibernate.show-sql= true
